pyproject.toml
[project]
name = "floocast"
version = "1.2.0"
description = "Configure and control FlooGoo USB Bluetooth dongles (FMA120)"
readme = "README.md"
license = {text = "BSD-3-Clause"}
requires-python = ">=3.10"
dependencies = [
    "pyserial>=3.5",
    "sounddevice",
    "numpy",
    "certifi",
    "pillow",
    "urllib3",
    "pystray",
    "notify-py",
]

[project.urls]
Homepage = "https://www.flairmesh.com/Dongle/FMA120.html"

[project.scripts]
floocast = "floocast.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/floocast"]

[tool.hatch.build.targets.sdist]
include = ["src/floocast"]

[project.optional-dependencies]
dev = [
    "ruff>=0.8",
    "pytest>=8.0",
    "pre-commit>=4.0",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",
    "F",
    "W",
    "I",
    "UP",
    "B",
    "C4",
    "SIM",
]
ignore = [
    "E501",
    "E701",
    "E702",
    "E711",
    "UP031",
    "SIM102",
    "SIM105",
    "SIM211",
    "F821",
    "F403",
    "F405",
    "F811",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
